+++ CHANGELOG.md
@@ -8,6 +8,7 @@ This CHANGELOG follows the format listed at [Keep A Changelog](http://keepachang
 ### Added
 - Add sensu go bonsai asset (@danragnar)
 - Bump bundler to 2.1
+- Allow to specify the LVM command to run
 
 ### Changed
 - appeased the cops (@majormoses)